Integrate one modal shared by rgbaSelector and colorSelector.
